[BUGFIX] Delete index documents without available site
Resolves: #3769
Releases: main, 11.5
Ports: #3770

+++ Classes/Domain/Index/Queue/GarbageRemover/AbstractStrategy.php
@@ -111,6 +111,10 @@ abstract class AbstractStrategy
             try {
                 $site = $indexQueueItem->getSite();
             } catch (InvalidArgumentException $e) {
+                $site = null;
+            }
+
+            if ($site === null) {
                 $this->queue->deleteItem($indexQueueItem->getType(), $indexQueueItem->getIndexQueueUid());
                 continue;
             }
